requirements.txt
numpy==1.23.2
Pillow==9.2.0
PyQt5==5.14.1
